fix: Id 컬럼을 Long에서 Int로 변경

+++ data/src/main/kotlin/com/sooum/data/repository/PlaceRepositoryImpl.kt
@@ -9,7 +9,7 @@ class PlaceRepositoryImpl@Inject constructor(
 
 ) : PlaceRepository {
 
-    override suspend fun addPlace(meetId: Long, shareResult: ShareResult) {
+    override suspend fun addPlace(meetId: Int, shareResult: ShareResult) {
 
     }
 }
